Uploaded Static MVC Project with Beta UI v1.0.0
This commit introduces our first version of the ClipperCoffeeCorner user interface. As of this commit, the UI files only consist of CSS and HTML files (containing <script> tags that should eventually be factored out). The webpage is static at the moment - currently no razer/partials/controllers/models - until we are ready to fully/properly implement them.